+++ tock/api/views.py
@@ -7,7 +7,6 @@ from django.db.models import Sum
 
 from django.contrib.auth.models import User
 from django.contrib.auth import get_user_model
-from django.contrib.auth.decorators import user_passes_test
 
 from projects.models import Project
 from hours.models import TimecardObject, Timecard, ReportingPeriod
@@ -47,7 +46,8 @@ class ProjectSerializer(serializers.ModelSerializer):
             'description',
             'billable',
             'start_date',
-            'end_date'
+            'end_date',
+            'active',
         )
     billable = serializers.BooleanField(source='accounting_code.billable')
     client = serializers.StringRelatedField(source='accounting_code')
@@ -119,21 +119,6 @@ class SlimBulkTimecardSerializer(serializers.Serializer):
     hours_spent = serializers.DecimalField(max_digits=5, decimal_places=2)
     billable = serializers.BooleanField(source='project.accounting_code.billable')
     mbnumber = serializers.CharField(source='project.mbnumber')
-
-class AdminBulkTimecardSerializer(serializers.Serializer):
-    project_name = serializers.CharField(source='project.name')
-    project_id = serializers.CharField(source='project.id')
-    employee = serializers.StringRelatedField(source='timecard.user')
-    start_date = serializers.DateField(source='timecard.reporting_period.start_date')
-    end_date = serializers.DateField(source='timecard.reporting_period.end_date')
-    hours_spent = serializers.DecimalField(max_digits=5, decimal_places=2)
-    billable = serializers.BooleanField(source='project.accounting_code.billable')
-    agency = serializers.CharField(source='project.accounting_code.agency.name')
-    flat_rate = serializers.BooleanField(source='project.accounting_code.flat_rate')
-    active = serializers.BooleanField(source='project.active')
-    mbnumber = serializers.CharField(source='project.mbnumber')
-    notes = serializers.CharField()
-    grade = serializers.CharField(source='grade.grade')
 
 # API Views
 
@@ -207,59 +192,6 @@ class TimecardList(generics.ListAPIView):
     def get_queryset(self):
         return get_timecards(self.queryset, self.request.query_params)
 
-def timeline_view(request, value_fields=(), **field_alias):
-    """ CSV endpoint for the project timeline viz """
-    queryset = get_timecards(TimecardList.queryset, request.GET)
-
-    fields = list(value_fields) + [
-        'timecard__reporting_period__start_date',
-        'timecard__reporting_period__end_date',
-        'project__accounting_code__billable'
-    ]
-
-    field_map = {
-        'timecard__reporting_period__start_date': 'start_date',
-        'timecard__reporting_period__end_date': 'end_date',
-        'project__accounting_code__billable': 'billable',
-    }
-    field_map.update(field_alias)
-
-    data = queryset.values(*fields).annotate(hours_spent=Sum('hours_spent'))
-
-    fields.append('hours_spent')
-
-    data = [
-        {
-            field_map.get(field, field): row.get(field)
-            for field in fields
-        }
-        for row in data
-    ]
-
-    response = HttpResponse(content_type='text/csv')
-
-    fieldnames = [field_map.get(field, field) for field in fields]
-    writer = csv.DictWriter(response, fieldnames=fieldnames)
-    writer.writeheader()
-    for row in data:
-        writer.writerow(row)
-    return response
-
-def project_timeline_view(request):
-    return timeline_view(
-        request,
-        value_fields=['project__id', 'project__name'],
-        project__id='project_id',
-        project__name='project_name',
-    )
-
-def user_timeline_view(request):
-    return timeline_view(
-        request,
-        value_fields=['timecard__user__username'],
-        timecard__user__username='user',
-    )
-
 def get_timecards(queryset, params=None):
     """
     Filter a TimecardObject queryset according to the provided GET
@@ -324,12 +256,6 @@ def slim_bulk_timecard_list(request):
     """
     queryset = get_timecards(TimecardList.queryset, request.GET)
     serializer = SlimBulkTimecardSerializer()
-    return stream_csv(queryset, serializer)
-
-@user_passes_test(lambda u: u.is_superuser)
-def admin_bulk_timecard_list(request):
-    queryset = get_timecards(TimecardList.queryset, request.GET)
-    serializer = AdminBulkTimecardSerializer()
     return stream_csv(queryset, serializer)
 
 from rest_framework.response import Response
