+++ tock/employees/models.py
@@ -1,5 +1,6 @@
 from django.db import models
 from django.contrib.auth.models import User
+from rest_framework.authtoken.models import Token
 
 class EmployeeGrade(models.Model):
     GRADE_CHOICES = (
@@ -73,3 +74,13 @@ class UserData(models.Model):
 
     def __str__(self):
         return '{0}'.format(self.user)
+
+    def save(self, *args, **kwargs):
+        if self.current_employee is False:
+            try:
+                token = Token.objects.get(user=self.user)
+                token.delete()
+            except Token.DoesNotExist:
+                pass
+
+        super(UserData, self).save(*args, **kwargs)
