+++ tock/tock/templates/base.html
@@ -49,7 +49,7 @@
         <li><b>Admin:</b></li>
         <li><a href="{% url 'reportingperiod:ReportingPeriodCreateView' %}">Add reporting period</a></li>
         <li><a href="/admin">Admin panel</a>
-        <li><a href="/api/timecards_bulk.csv">Bulk timecard CSV</a>
+        <li><a href="{% url 'reports:BulkTimecardList' %}">Bulk timecard CSV</a>
         {% endif %}
         <hr>
       </ul>
