+++ tock/tock/templates/hours/reports_list.html
@@ -7,8 +7,8 @@
 
 <h3> Raw data in .csv </h3>
 <ul>
-	<li><a href="http://tock.18f.gov/api/timecards_bulk.csv">Complete timecard data</a></li>
-	<li><a href="http://tock.18f.gov/api/slim_timecard_bulk.csv">Complete timecard data with fewer fields</a></li>
+	<li><a href="{% url 'reports:BulkTimecardList' %}">Complete timecard data</a></li>
+	<li><a href="{% url 'reports:SlimBulkTimecardList' %}">Complete timecard data with fewer fields</a></li>
 
 <div class="reporting-periods">
 	<h3>Reports by weekly reporting period</h3>
