+++ tock/tock/templates/projects/project_list.html
@@ -7,7 +7,7 @@
   <h3 class="chart__title">Project Hours Over Time</h3>
   <utilization-chart
     class="timeline timeline--projects"
-    data-url="/api/project_timeline.csv"
+    data-url="/reports/project_timeline.csv"
     href="/projects/{% verbatim %}{{ project_id }}{% endverbatim %}"
     layer="project_name"
     color="project"
